Add .gitignore, .prettierrc files and update README
These changes were done in order to improve our project's hygiene and readability. The .gitignore file was added to specify intentionally untracked files that Git should ignore. .prettierrc file was added to configure Prettier, which is an opinionated code formatter we're using. README file was updated to include detailed guidelines on coding challenges.