Refactor App to be a todo-list application
Unused components (Button.js and Time.js) were removed as they were not necessary in this app context. App.css was adjusted to better suit the new todo-list layout. The new TaskList, InputBar, and TaskItem components were created along with their respective css module files to structure and style the todo-list. A dependency, UUID, was added to handle unique IDs for each todo item.
App.js was cleaned up and restructured to process and render a todo-list, with handlers for adding and deleting todo tasks.