Refactor dummy data generation in shopping cart app
The dummyGenerator function responsible for creating dummy item data in the shopping app was refactored. The new version simplifies the data generation process by using the extended functionality of the Faker library directly, eliminating the need for additional dependency from 'uuid'. Also, the structure of the dummy data was modified to include priceCurrency and alt text for images, providing more complete and useful information for testing purposes.